various things

+++ scxrd/cif_model.py
@@ -93,6 +93,17 @@ class CifFileModel(models.Model):
             self.date_created = timezone.now()
         self.date_updated = timezone.now()
         super(CifFileModel, self).save(*args, **kwargs)
+        # Do not do this:
+        #print('Duplicates:', self.duplicates)
+        #for d in self.duplicates:
+        #    d.delete()
+
+    def find_duplicates(self):
+        return [i for i in CifFileModel.objects.exclude(pk=self.pk).filter(sha256=self.sha256)]
+
+    @property
+    def duplicates(self):
+        return self.find_duplicates()
 
     def __str__(self):
         try:
@@ -109,9 +120,11 @@ class CifFileModel(models.Model):
         return False
 
     def delete(self, *args, **kwargs):
+        if not self.exists:
+            return 
         cf = Path(self.cif_file_on_disk.path)
-        if DEBUG:
-            print('deleting', cf.name, 'in', cf.absolute())
+        #if DEBUG:
+        print('deleting', cf.name, 'in', cf.absolute())
         cf.unlink()
         super().delete(*args, **kwargs)
